feat: sign up data validation

+++ src/Pages/SignUp/index.jsx
@@ -1,11 +1,16 @@
-import { useContext, useRef } from 'react';
-import { Link } from 'react-router-dom';
+import { useContext, useRef, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { ShoppingCartContext } from '../../Context';
 import Layout from '../../Components/Layout';
 
 function SignUp() {
   const context = useContext(ShoppingCartContext);
   const form = useRef(null);
+  const [errorName, setErrorName] = useState('');
+  const [errorEmail, setErrorEmail] = useState('');
+  const [errorPassword, setErrorPassword] = useState('');
+  const [errorMessage, setErrorMessage] = useState('');
+  const navigate = useNavigate();
 
   // Account
   const account = localStorage.getItem('account');
@@ -17,19 +22,54 @@ function SignUp() {
     context.setSignOut(false);
   };
 
-  const createAnAccount = () => {
+  const createAnAccount = (event) => {
+    event.preventDefault();
+
     const formData = new FormData(form.current);
     const data = {
       name: formData.get('name'),
       email: formData.get('email'),
       password: formData.get('password'),
+      confirmPassword: formData.get('confirm-password'),
     };
-    // Create account
+
+    setErrorName('');
+    setErrorEmail('');
+    setErrorPassword('');
+    setErrorMessage('');
+
+    // Validar el nombre
+    if (!data.name) {
+      setErrorName('Please enter your name');
+      return;
+    }
+
+    // Validar el email
+    const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!data.email || !emailPattern.test(data.email)) {
+      setErrorEmail('Please enter a valid email address');
+      return;
+    }
+
+    // Validar la contraseña
+    if (!data.password || data.password.length < 6) {
+      setErrorPassword('Password must be at least 6 characters long');
+      return;
+    }
+
+    // Confirmar que las contraseñas coinciden
+    if (data.password !== data.confirmPassword) {
+      setErrorMessage("Passwords don't match");
+      return;
+    }
+
+    // Si todas las validaciones pasan, crear la cuenta
     const stringifiedAccount = JSON.stringify(data);
     localStorage.setItem('account', stringifiedAccount);
     context.setAccount(data);
     // Sign In
     handleSignIn();
+    navigate('/');
   };
 
   return (
@@ -43,33 +83,47 @@ function SignUp() {
             Create Account
           </h2>
           <form ref={form}>
-            <div className="mb-4">
+            <div className={errorName ? 'mb-2' : 'mb-4'}>
               <label className="block text-sm font-bold mb-2" htmlFor="name">
                 Name:
               </label>
               <input
-                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                className={`shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline  ${
+                  errorName && 'border-red-400'
+                }`}
                 id="name"
                 name="name"
                 type="text"
                 placeholder="Name"
-                defaultValue={parsedAccount?.name}
+                autoComplete="name"
               />
+              {errorName && (
+                <div className="text-center text-red-700 font-medium mt-2">
+                  {errorName}
+                </div>
+              )}
             </div>
-            <div className="mb-4">
+            <div className={errorEmail ? 'mb-2' : 'mb-4'}>
               <label className="block text-sm font-bold mb-2" htmlFor="email">
                 Email:
               </label>
               <input
-                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                className={`shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline  ${
+                  errorEmail && 'border-red-400'
+                }`}
                 id="email"
                 name="email"
                 type="text"
                 placeholder="Email"
-                defaultValue={parsedAccount?.email}
+                autoComplete="email"
               />
+              {errorEmail && (
+                <div className="text-center text-red-700 font-medium mt-2">
+                  {errorEmail}
+                </div>
+              )}
             </div>
-            <div className="mb-4">
+            <div className={errorPassword ? 'mb-2' : 'mb-4'}>
               <label
                 className="block text-sm font-bold mb-2"
                 htmlFor="password"
@@ -77,15 +131,23 @@ function SignUp() {
                 Password:
               </label>
               <input
-                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                className={`shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline  ${
+                  errorPassword && 'border-red-400'
+                }`}
                 id="password"
                 name="password"
                 type="password"
                 placeholder="********"
-                defaultValue={parsedAccount?.password}
+                autoComplete="new-password"
               />
+              {errorPassword && (
+                <div className="text-center text-red-700 font-medium mt-2">
+                  {errorPassword}
+                </div>
+              )}
             </div>
-            <div className="mb-6">
+
+            <div className={errorMessage ? 'mb-4' : 'mb-6'}>
               <label
                 className="block text-sm font-bold mb-2"
                 htmlFor="confirm-password"
@@ -93,23 +155,30 @@ function SignUp() {
                 Confirm Password:
               </label>
               <input
-                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 mb-3 leading-tight focus:outline-none focus:shadow-outline"
+                className={`shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline  ${
+                  errorMessage ? 'border-red-400 mb-2' : 'mb-4'
+                }`}
                 id="confirm-password"
                 name="confirm-password"
                 type="password"
                 placeholder="********"
+                autoComplete="new-password"
               />
+              {errorMessage && (
+                <div className="text-center text-red-700 font-medium">
+                  {errorMessage}
+                </div>
+              )}
             </div>
+
             <div className="text-center mb-2">
-              <Link to="/">
-                <button
-                  className="bg-gradient-to-r from-blue-500 to-blue-700 hover:from-blue-600 hover:to-blue-800 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-full"
-                  type="button"
-                  onClick={() => createAnAccount()}
-                >
-                  Create
-                </button>
-              </Link>
+              <button
+                className="bg-gradient-to-r from-blue-500 to-blue-700 hover:from-blue-600 hover:to-blue-800 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-full"
+                type="button"
+                onClick={(event) => createAnAccount(event)}
+              >
+                Create
+              </button>
             </div>
           </form>
         </div>
